derived quantity kernels
factor some of the derived quantities that were needed
to implement tc detection into the alg library. these
can now be resused and have been exposed via the python
binding as well.
